+++ src/app/popup/[popupId]/@sidebar/page.tsx
@@ -1,11 +1,11 @@
-import EventMapPanel from '@/features/event/detail/components/EventMapPanel';
-import { getEventById } from '@/features/event/detail/services/getEventById';
-import { getHostByEventId } from '@/features/event/detail/services/getHostByEventId';
-import type { PageProps } from '@/features/event/detail/types';
+import { getEventById } from '@/features/popup-detail/api/getEventById';
+import { getHostByEventId } from '@/features/popup-detail/api/getHostByEventId';
+import EventMapPanel from '@/features/popup-detail/components/EventMapPanel';
 import { getAddressPart } from '@/utils/adress';
 import { notFound } from 'next/navigation';
 
-export default async function Page({ params }: PageProps) {
+// export default async function Page({ params }: PageProps) {
+export default async function Page({ params }: { params: Promise<{ eventId: string }> }) {
   const { eventId } = await params;
   // 이벤트 조회
   const event = await getEventById(eventId);
@@ -17,6 +17,7 @@ export default async function Page({ params }: PageProps) {
   const place = getAddressPart(event.address!, 1);
 
   return (
+    // TODO: 이거 div 제거 요망
     <div className="mb-10 flex flex-col gap-4">
       <EventMapPanel address={place} organizer={host.name} />
     </div>
