--- src/features/README.md
@@ -1,1 +1,0 @@
-왜 안들어가
