+++ src/features/main/WrapperMain.tsx
@@ -1,15 +1,16 @@
+import { getMainPageData } from '@/features/main/api/getMainPageData';
 import MainBanner from '@/features/main/components/MainBanner';
 import OngoingPopupList from '@/features/main/components/OngoingPopupList';
 import UpcomingPopupList from '@/features/main/components/UpcomingPopupList';
-import { getMainPageData } from '@/features/main/services/getMainPageData';
 import { auth } from '@/libs/auth';
 
 export default async function WrapperMain() {
+  //
   const session = await auth();
   const { ongoingEvents, upcomingEvents, likedEventIds } = await getMainPageData(session?.user?.id);
 
   return (
-    <main className="w-100vw space-y-24 overflow-hidden">
+    <main className="space-y-24 overflow-hidden">
       <MainBanner />
       <OngoingPopupList sectionTitle="지금 인기있는 팝업!" events={ongoingEvents.events} likeEventIds={likedEventIds} />
       <UpcomingPopupList
