+++ src/features/main/api/getEvents.ts
