+++ src/features/main/api/getSavedStoreIds.ts
@@ -1,5 +1,6 @@
 import { clientApi } from '@/libs/api';
 
+// TODO: 네이밍 좀 좋은걸로
 export async function getSavedStoreIds(userId: string) {
   try {
     return await clientApi(`/api/users/${userId}/event-saves`, {
