+++ src/features/popup-detail/WrapperPopupDetail.tsx
@@ -1,10 +1,10 @@
 'use client';
 
-import EventDescription from '@/features/event/detail/components/EventDescription';
-import EventInfoList from '@/features/event/detail/components/EventInfoList';
-import EventMobileBar from '@/features/event/detail/components/EventMobileBar';
-import EventPoster from '@/features/event/detail/components/EventPoster';
-import EventTitle from '@/features/event/detail/components/EventTitle';
+import EventDescription from '@/features/popup-detail/components/EventDescription';
+import EventInfoList from '@/features/popup-detail/components/EventInfoList';
+import EventMobileBar from '@/features/popup-detail/components/EventMobileBar';
+import EventPoster from '@/features/popup-detail/components/EventPoster';
+import EventTitle from '@/features/popup-detail/components/EventTitle';
 import type { EventData } from '@/types/event';
 import type { User } from '@/types/user';
 import { Divider } from '@heroui/react';
@@ -15,10 +15,11 @@ interface EventDetailProps {
 }
 
 export default function WrapperPopupDetail({ event, host }: EventDetailProps) {
+  // TODO: 섬네일은 string값임 array 아님
   const imgSrc = Array.isArray(event.thumbnail) ? event.thumbnail[0] : event.thumbnail;
 
   return (
-    <>
+    <main>
       {/* 이미지 포스터(썸네일) */}
       <EventPoster src={imgSrc} alt={event.title} />
       {/* 타이틀 */}
@@ -37,6 +38,6 @@ export default function WrapperPopupDetail({ event, host }: EventDetailProps) {
       {/* 1024px 미만 반응형: 하단 바 */}
       {/* TODO: 관심토글 기능 구현해야함 */}
       <EventMobileBar priceLabel="무료" eventId={event.id} />
-    </>
+    </main>
   );
 }
