+++ src/features/popup-detail/api/getEventById.ts
@@ -6,6 +6,7 @@ export async function getEventById(id: string): Promise<EventData | null> {
     const data = await clientApi<EventData>(`/api/events/${id}`, { method: 'GET' });
     return data ?? null;
   } catch {
+    // TODO: 에러 처리 필요
     return null;
   }
 }
