+++ src/features/popup-detail/components/sidebar/.gitkeep
@@ -1,0 +1,1 @@
+여기에 사이드바 관련 컴포넌트, BL비즈니스 로직  옮기기
