--- src/features/popup-detail/types/index.ts
@@ -1,3 +1,0 @@
-export interface PageProps {
-  params: Promise<{ eventId: string }>;
-}
