+++ src/features/popup-register/WrapperPopupRegister.tsx
@@ -3,13 +3,13 @@
 import Button from '@/components/common/button';
 import Input from '@/components/common/input';
 import { LABELS } from '@/components/common/input/labels';
-import EventDatePicker from '@/features/event/register/components/EventDatePicker';
-import EventExtraInfo from '@/features/event/register/components/EventExtraInfo';
-import EventImageUpload from '@/features/event/register/components/EventImageUpload';
-import EventLocation from '@/features/event/register/components/EventLocation';
-import EventRecruitment from '@/features/event/register/components/EventRecruitment';
-import useEventRegisterForm from '@/features/event/register/hooks/useEventRegisterForm';
-import { OTHER_INFO_OPTIONS } from '@/features/event/register/services/otherInfoLabelHelper';
+import EventDatePicker from '@/features/popup-register/components/EventDatePicker';
+import EventExtraInfo from '@/features/popup-register/components/EventExtraInfo';
+import EventImageUpload from '@/features/popup-register/components/EventImageUpload';
+import EventLocation from '@/features/popup-register/components/EventLocation';
+import EventRecruitment from '@/features/popup-register/components/EventRecruitment';
+import useEventRegisterForm from '@/features/popup-register/hooks/useEventRegisterForm';
+import { OTHER_INFO_OPTIONS } from '@/features/popup-register/services/otherInfoLabelHelper';
 import { Textarea } from '@heroui/react';
 import { useSession } from 'next-auth/react';
 import Script from 'next/script';
