+++ src/features/popup-search/components/SelectableChip.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import Chip from '../chip';
+import Chip from '@/components/common/chip';
 
 interface Props {
   label: string;
@@ -8,7 +8,9 @@ interface Props {
   onClick: () => void;
 }
 
-export default function ClickableChip({ label, isSelected, onClick }: Props) {
+// NOTE: 태그 & 지역 필터링 칩 컴포넌트
+// TODO: 네이밍 변경 필요
+export default function SelectableChip({ label, isSelected, onClick }: Props) {
   return (
     <button type="button" onClick={onClick} className="focus:outline-none">
       <Chip size="lg" className={`cursor-pointer ${isSelected ? 'bg-blue-600 text-white hover:bg-blue-700' : ''}`}>
