+++ src/features/popup-search/services/fetchEvents.ts
@@ -10,6 +10,8 @@ export interface EventsResponse {
  * zone, page, pageSize 만 넘기면
  * 백엔드에서 받아온 { events, totalCount } 그대로 리턴
  */
+// TODO:네이밍 getEvents로 변경
+// api pages에도 쓰임 후에 따로 빼야될 듯?
 export async function fetchEvents(params: {
   zone?: string | null;
   page?: number;
